Added antialiasing options

+++ src/tacticshooter/Unit.java
@@ -303,10 +303,10 @@ public class Unit
 		}
 		
 		g.setColor( Color.black );
-		g.fillRect( -9, healthBarDist, (int)(18.f * health/type.health), 4 );
+		g.fillRect( -9, healthBarDist, (18.f * health/type.health), 4 );
 		
 		g.setColor( new Color( DMath.bound( 1.f - health/type.health, 0, 1 ), DMath.bound(health/type.health, 0, 1 ), 0 ) );
-		g.fillRect( -8, healthBarDist+1, (int)(16.f * health/type.health), 2 );
+		g.fillRect( -8, healthBarDist+1, (16.f * health/type.health), 2 );
 		
 		float dmx = x - mx;
 		float dmy = y - my;
